Update LS7366R.cpp
Very nice library here. I found that the MDR0, MDR1 configuration changes were not taking effect on the right encoder. To fix that, I set the chip select states prior to the first SPI call. (After pinMode() is called, the state of the pin is set to what it was when last set as an output pin but here we don't know the pin's prior state.)